Handle dist info casing mismatch in worker
The metadata name may be uppercase, while the wheel and dist info names are lowercase, or
the metadata name and the dist info name are lowercase, while the wheel name is uppercase.
Either way, we just search the wheel for the name. See `find_dist_info`:
https://github.com/astral-sh/puffin/blob/2652caa3e31282afc2f1e1ca581ac4f553af710d/crates/install-wheel-rs/src/wheel.rs#L1024-L1057

I tested this with `wrangler dev` and `bio_embeddings[all]`