distribution-filename: add note about dropping build tags
I added this comment because I was trying to figure out where and
how the build tags were used before modifying the code. But it looks
like they aren't. I also checked and, at least the `flyte.in`
benchmark does have some wheel names that contain a build number.
There is perhaps a latent bug here, but I'm not certain.