bench: add micro-benchmarks for wheel-related operations
The goal here is to optimize WheelFilename::is_compatible. In so doing,
we want to not only track our progress to ensure a real improvement has
been made, but also that we don't regress other things. (Or, if we do,
that such regressions are okay.) For example, speeding up
WheelFilename::is_compatible will likely require making Tags::new
slower. Which is probably fine so long as it remains "reasonable" since
Tags::new is only executed ~once for the lifetime of a `puffin` process.

The benchmarking structure here was mostly copied from ruff_benchmark.