distribution-filename: add tests for WheelFilename parsing
Before tweaking this to improve the speed of tag compatibility checks,
we add a few tests to check basic parsing behavior. We try to cover all
error conditions.

In the process of doing this, I changed one error condition to an unwrap
since it's impossible for str::split to return an empty iterator for any
inputs.

This also adds Insta as a dependency of distribution-filename for the
tests.