distribution-filename: reject wheel filenames with too many components
Previously, the wheel filename parser would accept names with more than
6 components. It would behave as if any component after the sixth did
not exist.

PEP 0427 seems to imply that there really can be no more than 6
components. In particular, if more were allowed or if dashes were
allowed in the platform tag, then it would be ambiguous in some
cases whether a wheel filename had a build number or not.

The reason for adding this error case is because we want to depend
on the fact that none of the tags contain dashes when optimizing.
Technically we could do so without this error case, but without it,
the code doesn't quite make it clear whether anything extra is left
out by mistake or not. By adding an explicit error case, we make it
a little harder to accidentally break assumptions made by other
parts of the code.