distribution-filename: speed up is_compatible
This commit tweaks the representation of `Tags` in order to offer a
faster implementation of `WheelFilename::is_compatible`. We now use a
nested map of tags that lets us avoid looping over every supported
platform tag. As the code comments suggest, that is the essential gain.
We still do not mind looping over the tags in each wheel name since they
tend to be quite small. And pushing our thumb on that side of things can
make things worse overall since it would likely slow down WheelFilename
construction itself.

For micro-benchmarks, we improve considerably for compatibility
checking:

    $ critcmp base test3
    group                                                   base                                    test3
    -----                                                   ----                                    -----
    build_platform_tags/burntsushi-archlinux                1.00     46.2±0.28µs        ? ?/sec     2.48    114.8±0.45µs        ? ?/sec
    wheelname_parsing/flyte-long-compatible                 1.00    624.8±3.31ns   174.0 MB/sec     1.01    629.4±4.30ns   172.7 MB/sec
    wheelname_parsing/flyte-long-incompatible               1.00    743.6±4.23ns   165.4 MB/sec     1.00    746.9±4.62ns   164.7 MB/sec
    wheelname_parsing/flyte-short-compatible                1.00    526.7±4.76ns    54.3 MB/sec     1.01    530.2±5.81ns    54.0 MB/sec
    wheelname_parsing/flyte-short-incompatible              1.00    540.4±4.93ns    60.0 MB/sec     1.01    545.7±5.31ns    59.4 MB/sec
    wheelname_parsing_failure/flyte-long-extension          1.00     13.6±0.13ns     3.2 GB/sec     1.01     13.7±0.14ns     3.2 GB/sec
    wheelname_parsing_failure/flyte-short-extension         1.00     14.0±0.20ns  1160.4 MB/sec     1.01     14.1±0.14ns  1146.5 MB/sec
    wheelname_tag_compatibility/flyte-long-compatible       11.33   159.8±2.79ns   680.5 MB/sec     1.00     14.1±0.23ns     7.5 GB/sec
    wheelname_tag_compatibility/flyte-long-incompatible     237.60 1671.8±37.99ns    73.6 MB/sec    1.00      7.0±0.08ns    17.1 GB/sec
    wheelname_tag_compatibility/flyte-short-compatible      16.07   223.5±8.60ns   128.0 MB/sec     1.00     13.9±0.30ns     2.0 GB/sec
    wheelname_tag_compatibility/flyte-short-incompatible    149.83   628.3±2.13ns    51.6 MB/sec    1.00      4.2±0.10ns     7.6 GB/sec

We do regress slightly on the time it takes for `Tags::new` to run, but
this is somewhat expected. And in absolute terms, 114us is perfectly
acceptable given that it's only executed ~once for each `puffin`
invocation.

Ad hoc benchmarks indicate an overall 25% perf improvement in `puffin
pip-compile` times. This roughly corresponds with how much time
`is_compatible` was taking. Indeed, a profile confirm that it has
virtually disappeared from the profile.

Fixes #157