pep508-rs: make trailing whitespace explicit
My editor is configured to remove all trailing whitespace on every line
whenever I save. I find that configuration to be extremely useful and
would prefer not to turn it off. Unfortunately, the tests in this file
do rely on significant trailing whitespace that was getting stripped
whenever I saved the file. I've tweaked the strings to avoid this by
inserting explicit an explicit ASCII whitespace escape (\x20). Not
great, but if we only do this rarely it seems okay?