pep440: rejigger imports
The imports seemed like they had organically become a bit jumbled, so
I've straightened them out here. I don't mean to start a crusade for
this (especially since rustfmt doesn't have support for doing this
automatically yet on stable Rust), but I felt like it was jumbled enough
to be worth getting a refresh.

Note that in some cases I removed imports in favor of fully qualified
names. I usually do this for names that aren't used too frequently or
are otherwise short enough. This is more of a style thing.