pep440: make Version an opaque type
The idea here is to make it easier to iterate on how a Version
is represented internally. When its representation is exposed, this
is much harder.

We don't actually change the representation here, so the change is
pretty straight-forward: we migrate to new accessors and mutators.

One thing we choose not to do (currently) is expose any way of
mutating a version in place. There's no specific reason for this
at the moment, but that 1) it didn't seem like anything really
needed it and 2) not needing to support in-place mutation may
make it more flexible to experiment with its internals.