pep440: rejigger without_local and drop from_release
This makes `without_local` take `self` like the other mutators, and by
doing so, avoids the required clone. (Although it will force the caller
to clone explicitly if they have a `&Version`. Which is the case here.
The number of absolute clones doesn't change, just where they are. But
in theory, it could be reduced if the caller had a `Version`.) I didn't
do this in a prior commit because I missed it.

This also gets rid of `from_release` and uses the `new` constructor that
accepts any iterator of release numbers. While being able to reuse the
`Vec` given by the caller is perhaps an advantage currently, this may
not always be the case. On top of that, it was rarely used in Puffin and
the caller was allocating a new `Vec` anyway.