pep440: add regression test for non-ASCII version specifier
It turns out that the current parser permits non-ASCII whitespace in
places. We can be sneaky here to cause the existing implementation to
produce incorrect offsets.

The core issue here is that `unicode-width` is used to compute codepoint
offsets, but its actual purpose is to compute the *visual width* of a
codepoint that has been rendered. Some codepoints use more than 1 unit
of visual width.

While this establishes a mismatch between the implementation and the
documented behavior of `Pep440Error`, we will also address this by
switching to byte offsets instead of codepoint offsets. (Codepoint
offsets are almost never what you want.)