pep440: rejigger error type
This error type was only be used when one attempted to parse a string as
a `VersionSpecifiers`. We'll want to introduce more structured error
types for parsing `Version` and `VersionSpecifier` as well, so renaming
the error type helps make room for that.

We also make the error type opaque. Nothing (in puffin at least) seemed
to need its internals. We can always add accessor methods in the future
if something else needs it. It's overall rare to need to expose the
entire internal representation of an error type.