pep440: use structured errors for version specifiers
This smoothes out the error handling to using something a bit more
structured. I did this because I intend to do the same for Version
(eventually), and it seems good to be consistent. It also lets us nest
errors a bit more easily and scrutinize the different error classes at a
glance.