pep440: add VersionPattern
This starts to untangle the dual use of Version for both acting as a
literal version and a pattern for matching versions. We still retain
reuse in the implementation, but are more explicit about the difference
in the public API. This also makes it a little smoother to build a
`VersionSpecifier` instead of carrying an extra boolean around.