pep440: add "fast" path for version parsing
This adds a simple fast path to version parsing that just looks for
versions that only have a release component with 4 or fewer numbers.

This does lead to a small ~1.03x improvement on the benchmark mentioned
in a previous commit, but the main reason to do this to set ourselves up
for version parsing without any allocation at all. Namely, in the next
commit, we're going to split the representation of Version in two.