pep440: split Version into "small" and "full" components
The high level of this commit is to make `Version::cmp` lower to the
equivalent of `u64::cmp` in the vast majority of cases.

We achieve this by carving out a large subset (~90%) of all possible
versions on PyPI and devising a small fixed representation for them.
This is called `VersionSmall`. Everything that doesn't fit into this
representation is put into `VersionFull`. The former is, essentially, a
`u64`. The latter is what the status quo looked like before this commit.

This change overall makes a rather large difference when puffin gets
stuck in version resolution. Namely, pubgrub does a lot of version
comparisons, and by changing that a simple u64 comparison, we get a huge
win:

    $ time puffin-cmp pip-compile --no-build --cache-dir ~/astral/tmp/cache/ -o /dev/null ./scripts/requirements/boto3.in
    Resolved 31 packages in 10.90s

    real    10.923
    user    10.406
    sys     0.384
    maxmem  2398 MB
    faults  0

    $ time puffin-base pip-compile --no-build --cache-dir ~/astral/tmp/cache/ -o /dev/null ./scripts/requirements/boto3.in
    Resolved 31 packages in 34.56s

    real    34.579
    user    34.004
    sys     0.413
    maxmem  2867 MB
    faults  0

Future work on this point might investigate removing some of the extra
storage on `VersionSmall` (see code comments). And if we did that, it
might make sense to box a `Version`.